remove setting allowFullPageRender
It seems to me that this is not required anymore. I couldn’t find this property in React’s or ReactMount’s current source code and it doesn’t break anything. There also seems to be no documentation about this property or I couldn’t find it. Please let me know if I'm mistaken.